requirements.txt
streamlit
scikit-learn
xgboost
joblib
pandas
numpy
Pillow
shap
matplotlib

